stories/api/serializers.py
from rest_framework import serializers
from api.models import *


class AnswerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Answer
        fields = ('text', 'nextQuestion', 'anim', 'damage', 'question')


class QuestionSerializer(serializers.ModelSerializer):
    answers = AnswerSerializer(many=True, read_only=True)

    class Meta:
        model = Question
        fields = ('prompt', 'answers', 'story')


class StorySerializer(serializers.ModelSerializer):
    questions = QuestionSerializer(many=True, read_only=True)

    class Meta:
        model = Story
        fields = ('title', 'questions', 'description')
